Applied naming standard to files
