+++ tests/GoatTest.py
